feature/new-news main news

+++ app/src/main/java/com/maxim/diaryforstudents/core/service/CloudNews.kt
@@ -3,6 +3,6 @@ package com.maxim.diaryforstudents.core.service
 data class CloudNews(
     val title: String = "",
     val content: String = "",
-    val date: Int = 0,
+    val date: Long = 0L,
     val photoUrl: String = ""
 )
